pyproject.toml
[project]
name = "bza-tool"
version = "0.1.0"
description = "Benchmark how quantization affects retention of ROME-implanted facts in LLMs"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "datasets",
    "accelerate",
    "auto-gptq",
    "llmcompressor",
    "pyyaml",
    "tqdm",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
